# specs/002-backend-setup-fastapi/contracts/todo-api-openapi.yaml
openapi: 3.0.0
info:
  title: Todo Backend API
  version: 1.0.0
  description: API for managing user tasks with authentication and user isolation

servers:
  - url: http://localhost:8000
    description: Development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Title of the task
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000
          nullable: true
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        user_id:
          type: string
          description: ID of the user who owns the task
          example: "user-12345"
        created_at:
          type: string
          format: date-time
          description: When the task was created
          example: "2023-01-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated
          example: "2023-01-01T12:00:00Z"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the task
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000
          nullable: true
          example: "Milk, eggs, bread"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the task
          maxLength: 255
          example: "Buy groceries and cook dinner"
        description:
          type: string
          description: Updated description of the task
          maxLength: 1000
          nullable: true
          example: "Milk, eggs, bread, chicken"
        completed:
          type: boolean
          description: Updated completion status
          example: true

    TaskCompletionUpdate:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: New completion status
          example: true

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: Health status of the service
          example: "healthy"

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Not found"

security:
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the service
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  
  /api/tasks:
    get:
      summary: Get all tasks for the authenticated user
      description: Retrieves all tasks associated with the authenticated user
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieves a specific task by ID for the authenticated user
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      summary: Update a task
      description: Updates a specific task by ID for the authenticated user
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Delete a task
      description: Deletes a specific task by ID for the authenticated user
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tasks/{id}/complete:
    patch:
      summary: Update task completion status
      description: Updates the completion status of a specific task by ID for the authenticated user
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompletionUpdate'
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'